Extract GenericTypesCollector into a dedicated class
CBridge generator and Dart generator each had a dedicated
private class for collecting generic type references.
Extracted these into a single dedicated class
GenericTypesCollector.
