+++ gluecodium/src/test/java/com/here/gluecodium/SmokeTest.kt
@@ -28,7 +28,6 @@ import org.junit.runners.Parameterized
 import org.junit.runners.Parameterized.Parameters
 
 @RunWith(Parameterized::class)
-
 class SmokeTest(
     private val featureDirectory: File,
     generatorName: String,
